+++ source/nanoFramework.Hardware.Stm32/Power.WakeupPin.cs
@@ -1,0 +1,31 @@
+﻿//
+// Copyright (c) 2018 The nanoFramework project contributors
+// See LICENSE file in the project root for full license information.
+//
+
+namespace nanoFramework.Hardware.Stm32
+{
+    public partial class Power
+    {
+        /// <summary>
+        /// Wake-up pin enumeration.
+        /// </summary>
+        public enum WakeupPin
+        {
+            /// <summary>
+            /// Wake-up pin 1.
+            /// </summary>
+            Pin1 = 1,
+
+            /// <summary>
+            /// Wake-up pin 2.
+            /// </summary>
+            Pin2,
+
+            /// <summary>
+            /// Wake-up pin 3.
+            /// </summary>
+            Pin3
+        }
+    }
+}